refactor SecurityProfitService and AssetsAndCashService
